add Burger menu
залишилось адаптувати  бургер меню,та виправити стилі
