+++ src/components/SharedLayout/ChangeHeadingName/ChangeHeadingName.jsx
@@ -1,0 +1,6 @@
+const { Heading } = require('./ChangeHeadingName.styled');
+
+const ChangeHeadingName = () => {
+  return <Heading>Calendar</Heading>;
+};
+export default ChangeHeadingName;
