+++ src/components/SharedLayout/ChangeHeadingName/ChangeHeadingName.styled.jsx
@@ -1,0 +1,17 @@
+import styled from 'styled-components';
+export const Heading = styled.h1`
+  display: none;
+
+  @media screen and (min-width: 1440px) {
+    display: flex;
+
+    color: #111111;
+    text-shadow: 0px 9.399999618530273px 57.6875px rgba(0, 0, 0, 0.04),
+      0px 47px 355px rgba(0, 0, 0, 0.07);
+    font-family: Inter;
+    font-size: 32px;
+    font-style: normal;
+    font-weight: 700;
+    line-height: 1;
+  }
+`;
