+++ src/components/SharedLayout/Header/Header.jsx
@@ -1,0 +1,11 @@
+import ChangeHeadingName from '../ChangeHeadingName/ChangeHeadingName';
+import { TopMenuLeftPart } from './Header.styled';
+const Header = () => {
+  return (
+    <TopMenuLeftPart>
+      
+      <ChangeHeadingName />
+    </TopMenuLeftPart>
+  );
+};
+export default Header;
