+++ src/components/SharedLayout/SharedLayout.jsx
@@ -1,97 +1,40 @@
 import { Outlet } from 'react-router-dom';
-import React from 'react';
+import React, { useState } from 'react';
 
 import {
   Main,
   Container,
-  LeftDesktopMenu,
   MainContent,
-  TopMenuLeftPart,
   TopMenu,
-  BurgerMenuButton,
-  Heading,
-  ProfInfoWrap,
-  FeedbackButton,
-  ThemeToggleButton,
-  NameField,
-  AvatarThumb,
-  Avatar,
-  LogoBlock,
-  LogoImg,
-  LogoText,
-  LeftMenuTitle,
-  LeftMenuItemIco,
-  LeftMenuItemLink,
-  LeftMenuItemText,
-  LeftMenuLogOutButton,
-  LogOutIcoEl,
+  BurgerSide,
 } from './SharedLayout.styled';
 
+import SiteNav from './SiteNav/SiteNav';
+import UserInfo from './UserInfo/UserInfo';
+import Header from './Header/Header';
 import BurgerMenuIce from '../../images/burger-menu-01.svg';
-import ThemeIcoButton from '../../images/theme-ico-button-moon.svg';
-import UserAvatar from '../../images/user-avatar.png';
-
-import LogoGoose from '../../images/logo-goose-left-menu.png';
-
-import AccountIco from '../../images/user-check-01.svg';
-import CalendarIco from '../../images/calendar-check-02.svg';
-import StatisticsIco from '../../images/chart-2.svg';
-
-import LogOutIco from '../../images/log-out-01.svg';
+import { BurgerMenuButton } from './Header/Header.styled';
 
 const SharedLayout = () => {
+  const [isMenuOpen, setIsMenuOpen] = useState(false);
+  const handleToggle = () => setIsMenuOpen(pS => !pS);
   return (
     <Main>
       <Container>
-        {/* LEFT-CONTENT */}
-        <LeftDesktopMenu>
-          <LogoBlock>
-            <LogoImg src={LogoGoose} alt="logo goose" />
-            <LogoText>GooseTrack</LogoText>
-          </LogoBlock>
-          <LeftMenuTitle>User Panel</LeftMenuTitle>
-
-          <LeftMenuItemLink to="/account">
-            <LeftMenuItemIco src={AccountIco} />
-            <LeftMenuItemText>My account</LeftMenuItemText>
-          </LeftMenuItemLink>
-
-          <LeftMenuItemLink to="/calendar">
-            <LeftMenuItemIco src={CalendarIco} />
-            <LeftMenuItemText>Calendar</LeftMenuItemText>
-          </LeftMenuItemLink>
-
-          <LeftMenuItemLink to="/statistics">
-            <LeftMenuItemIco src={StatisticsIco} />
-            <LeftMenuItemText>Statistics</LeftMenuItemText>
-          </LeftMenuItemLink>
-
-          <LeftMenuLogOutButton>
-            Log out
-            <LogOutIcoEl src={LogOutIco} alt="logout button" />
-          </LeftMenuLogOutButton>
-        </LeftDesktopMenu>
-        {/* RIGHT-CONTENT */}
+        <BurgerSide data={isMenuOpen ? 'true' : undefined}>
+          <SiteNav onClose={handleToggle} isOpen={isMenuOpen} />
+        </BurgerSide>
         <MainContent>
           <TopMenu>
-            <TopMenuLeftPart>
-              <BurgerMenuButton type="button">
+            {isMenuOpen ? (
+              null
+            ) : (
+              <BurgerMenuButton type="button" onClick={handleToggle}>
                 <img src={BurgerMenuIce} alt="open menu" width="100%" />
               </BurgerMenuButton>
-
-              <Heading>Calendar</Heading>
-            </TopMenuLeftPart>
-
-            <ProfInfoWrap>
-              <FeedbackButton type="button">Feedback</FeedbackButton>
-              <ThemeToggleButton type="button">
-                <img src={ThemeIcoButton} alt="change theme" width="100%" />
-              </ThemeToggleButton>
-              <NameField>Nadiia</NameField>
-              <AvatarThumb>
-                <Avatar src={UserAvatar} alt="user avatar" />
-              </AvatarThumb>
-            </ProfInfoWrap>
+            )}
+            <Header />
+            <UserInfo />
           </TopMenu>
 
           <Outlet />
